Tidy-up bitset implementation

+++ 2022/10/day10_impl.h
@@ -4,12 +4,10 @@
 #include <stdlib.h>
 #include <string.h>
 #include "file_loading.h"
-#include "set_tuple2.h"
 #include "util.h"
-#include "bitset.h"
 
-static const int HEIGHT = 6;
-static const int WIDTH = 40;
-static const int SPRITE_WIDTH = 3;
+static const size_t HEIGHT = 6;
+static const size_t WIDTH = 40;
+static const size_t SPRITE_WIDTH = 3;
 
 void calculate(char** data, uint64_t data_size, int64_t* part1, char* part2);
